requirements.txt
yfinance>=0.2.31
mplfinance>=0.12.10b0
google-genai>=1.0.0
requests>=2.31.0
Pillow>=10.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
